Tentatively remove limit file
Sometimes I would get rm file not found errors.
Since everything else works as expected, it is
quicker to do this than properly debug it.

+++ .auto_build_template.sh
@@ -24,4 +24,6 @@ if [ "$LIMIT" -nt "$LAST" ]; then
     touch "$LAST"
 fi
 
-rm "$LIMIT"
+if [[ -f "$LIMIT" ]] ; then
+    rm "$LIMIT"
+fi
